requirements.txt
psycopg2-binary
sqlalchemy
plotly
pandas
python-dotenv
streamlit
instructor[google-generativeai]